[utils & Kernel utils]: moved most stuff from kernel::utils to a separtare crate utils
