+++ kernel/Cargo.toml
@@ -5,10 +5,9 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-noto-sans-mono-bitmap = { version = "0.3.0", features = [
-    "size_20",
-] }
+noto-sans-mono-bitmap = { version = "0.3.0", features = ["size_20"] }
 
+safa-utils = { path = "../utils" }
 lazy_static = { version = "1.5.0", features = ["spin_no_std"] }
 spin = "0.9.8"
 bitflags = "2.6.0"
@@ -19,7 +18,9 @@ limine = "0.3.0"
 hashbrown = "0.15.2"
 
 serde = { version = "1.0", default-features = false, features = ["derive"] }
-serde_json = { version = "1.0", git = "https://github.com/ObserverUnit/srede-json-frok", default-features = false, features = ["alloc"] }
+serde_json = { version = "1.0", git = "https://github.com/ObserverUnit/srede-json-frok", default-features = false, features = [
+    "alloc",
+] }
 
 compile-time = "0.2"
 slab = { version = "0.4", default-features = false }
