+++ utils/src/bstr.rs
